divide returnerar nu en double

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -7,7 +7,7 @@ public class MyMathCalc {
     public int multiply(int a, int b) {
         return a*b;
     }
-    public int divide(int a, int b) {
+    public double divide(double a, double b) {
         return a/b;
     }
 }
